--- packages/create-cloudflare/templates-experimental/svelte/templates/static/.assetsignore
@@ -1,4 +1,0 @@
-_worker.js
-_routes.json
-_headers
-_redirects
