+++ packages/create-cloudflare/templates/svelte/pages/templates/wrangler.jsonc
