+++ packages/create-cloudflare/templates/svelte/workers/c3.ts
@@ -7,7 +7,7 @@ import { usesTypescript } from "helpers/files";
 import { detectPackageManager } from "helpers/packageManagers";
 import { installPackages } from "helpers/packages";
 import * as recast from "recast";
-import type { TemplateConfig } from "../../src/templates";
+import type { TemplateConfig } from "../../../src/templates";
 import type { C3Context, PackageJson } from "types";
 
 const { npm } = detectPackageManager();
@@ -23,7 +23,7 @@ const configure = async (ctx: C3Context) => {
 	const pkg = `@sveltejs/adapter-cloudflare`;
 	await installPackages([pkg], {
 		dev: true,
-		startText: "Adding the Cloudflare Pages adapter",
+		startText: "Adding the Cloudflare adapter",
 		doneText: `${brandColor(`installed`)} ${dim(pkg)}`,
 	});
 
@@ -103,7 +103,7 @@ const config: TemplateConfig = {
 	copyFiles: {
 		path: "./templates",
 	},
-	path: "templates-experimental/svelte",
+	path: "templates/svelte/workers",
 	generate,
 	configure,
 	transformPackageJson: async (original: PackageJson, ctx: C3Context) => {
