+++ packages/create-cloudflare/templates/svelte/workers/templates/static/.assetsignore
@@ -1,0 +1,2 @@
+_worker.js
+_routes.json
