+++ packages/create-cloudflare/templates/svelte/workers/templates/wrangler.jsonc
